backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
pymysql==1.1.0

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2

# HTTP client
httpx==0.26.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# LLM SDKs
openai==1.10.0

# Elasticsearch
elasticsearch[async]==8.12.0

# Utilities
python-dotenv==1.0.0
